stt/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper>=20231117
